feat: add available-routes api endpoint (#219)
* improve: export getSpokePool

* feat: add route availability endpoint

* improve: add param filters

* improve: increase cache duration

* fix: update docs to reflect new cache duration

* test: include additional mocked functions

* test: add unit test for available routes

* fix: change additional docs to reflect new cache

* improve: prefer json data instead of on-chain data

* fix: correct error to account for checksum case

* fix: remove lowercase string transformations

* improve: remove requirement for on-chain data

* fix: make comparisons case insensitive

* fix: lower time complexity of destination lookup

* fix: change variable to const

* fix: prefer const to let

Co-authored-by: Matt Rice <matthewcrice32@gmail.com>

* fix(api): further improve runtime complexity

This change improves performance by reducing an O(n) loop. Additionally, by rearranging the code structure, a more succint map/filter structure is defined in which the chained values resolve to a single variable

* perf(api): prefer filter before map

By definition, the filter will produce an output array that will be less than or equal to the length of the origin array. Filtering this first means that the mapping may need to be applied over less elements

* perf(api): perform a filter-map to collapse two loops into one

This change introduces a filterMap utility function that allows an array to be filtered & mapped in one pass using a reduce function. This is applied to the enabledRoutes to filter against query parameters & map the route to the original return type of the API endpoint.

* perf(api): discern a single if statement prior to reduce

This change hoists the if statement check if `firstMap` is chosen to before the reduce function. The intuition behind this change is that we now only perform one if statement rather than computing it for each iteration

Co-authored-by: Matt Rice <matthewcrice32@gmail.com>