improve(limits): Dynamically source LP cushion (#492)
A follow-up commit needs to map symbol, address and decimals, so it
makes sense to factor it out. As part of this work, it became obvious that
we can simplify the lookup of the LP cushion and cut out a lot of code.

For speed, require that addresses are checksummed to avoid the need to
normalise them on each lookup.

Follow-up work will factor out the token definitions to a common source.