fix(launch): :bug: prevent duplicate http calls to details API (#521)
This change performs two functional modifications to the `useQuery` call when in the referrals page. Firstly, a staleTime limit is imposed such that referral details data will wait a period of 10 seconds before refreshing the stale client-side data with the server data. Secondly, this change enforces that the inner logic of the specific query is using the current query key.

Breaking Changes: N/A