fix(amplitude): :necktie: correct timestamp in amplitude telemetry (#603)
This PR proposes updating the latency timestamps in Amplitude to improve the accuracy of the data feed. The current timestamps are not taking into account the time it takes for the data to be processed, which can lead to inaccuracies in the data feed. By updating the timestamps to include this processing time, we can ensure that the data feed is as accurate as possible.