feat: 🔥 remove boba chain from FE (#676)
* refactor: :fire: remove boba from FE

This change effectively removes the Boba Network/ Boba Token from the front end. Boba will no longer appear in any copy or in any routes that are enabled by Across. This is enforced at both the front end and in the contracts. No RPC calls to the Boba network will be made.

* feat: :bug: enable boba for only pools

This change enables BOBA for pools to allow users to offboard. This required additional logic changes because the primary logic to display pools was within the data json files. Since we removed all BOBA routes within the json, additional structures needed to be created.

* fix: revert removal of certain segments

* chore: refactor

* fix: remove errant log